en2ar-service/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
transformers==4.36.2
torch==2.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentencepiece==0.1.99
gunicorn==21.2.0
